pyproject.toml
[project]
name = "kaggle-titanic"
version = "0.1.0"
description = ""
authors = [
    {name = "Vlad Mashtalier",email = "vmashtaler5@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "contourpy (==1.3.0)",
    "cycler (==0.12.1)",
    "fonttools (==4.54.0)",
    "joblib (==1.4.2)",
    "kiwisolver (==1.4.7)",
    "matplotlib (==3.9.2)",
    "numpy (==2.1.0)",
    "packaging (==24.1)",
    "pandas (==2.2.2)",
    "pillow (==10.4.0)",
    "pyparsing (==3.1.4)",
    "python-dateutil (==2.9.0.post0)",
    "pytz (==2024.1)",
    "scikit-learn (==1.5.1)",
    "scipy (==1.14.0)",
    "six (==1.16.0)",
    "threadpoolctl (==3.5.0)",
    "tzdata (==2024.1)",
    "xgboost (==2.1.1)",
    "mlflow (>=2.22.0,<3.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "optuna (>=4.3.0,<5.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"